requirements.txt
aiohttp
feedparser
apscheduler
beautifulsoup4
python-dotenv
